finish refactor to player group
