+++ music_assistant/client/players.py
@@ -174,10 +174,6 @@ class Players:
             "players/cmd/group_volume", player_id=player_id, volume_level=volume_level
         )
 
-    async def set_player_group_power(self, player_id: str, power: bool) -> None:
-        """Handle power command for a (Sync)Group."""
-        await self.client.send_command("players/cmd/group_volume", player_id=player_id, power=power)
-
     async def set_player_group_members(self, player_id: str, members: list[str]) -> None:
         """
         Update the memberlist of the given PlayerGroup.
