+++ music_assistant/server/models/provider.py
@@ -101,6 +101,7 @@ class Provider:
             "domain": self.domain,
             "name": self.config.name or self.name,
             "instance_id": self.instance_id,
+            "lookup_key": self.lookup_key,
             "supported_features": [x.value for x in self.supported_features],
             "available": self.available,
             "is_streaming_provider": getattr(self, "is_streaming_provider", None),
