+++ music_assistant/server/providers/player_group/manifest.json
@@ -1,0 +1,13 @@
+{
+  "type": "player",
+  "domain": "player_group",
+  "name": "Playergroup",
+  "description": "Create (permanent) groups of your favorite players. \nSupports both syncgroups (to group speakers of the same ecocystem to play in sync) and universal groups to group speakers of different ecosystems to play the same audio (but not in sync).",
+  "codeowners": ["@music-assistant"],
+  "requirements": [],
+  "documentation": "https://music-assistant.io/player-support/player_group/",
+  "multi_instance": false,
+  "builtin": true,
+  "allow_disable": false,
+  "icon": "speaker-multiple"
+}
