--- music_assistant/server/providers/ugp/manifest.json
@@ -1,14 +1,0 @@
-{
-  "type": "player",
-  "domain": "ugp",
-  "name": "Universal Group Player",
-  "description": "Create Player Groups with your favorite players, regardless of type and model.",
-  "codeowners": [
-    "@music-assistant"
-  ],
-  "requirements": [],
-  "documentation": "https://music-assistant.io/player-support/universal/",
-  "multi_instance": false,
-  "builtin": false,
-  "icon": "speaker-multiple"
-}
